requirements.txt
datefinder==0.7.3
Flask==2.3.2
Flask_Cors==4.0.0
pdfplumber==0.10.2
pymongo==4.6.3
PyPDF2==3.0.1
python_dateutil==2.8.2
openai==0.27.0
python-dotenv==0.19.1
Werkzeug==2.3.6